fix: incorporate feedback
