--- src/test/resources/examples/differentErrorResponseBody/api.yaml
@@ -1,32 +1,0 @@
-openapi: 3.0.0
-info:
-  version: '1'
-  title: differentErrorResponseBody
-paths:
-  /example:
-    get:
-      responses:
-        200:
-          description: success
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/SuccessResponse"
-        400:
-          description: bad request
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/ErrorResponse"
-components:
-  schemas:
-    SuccessResponse:
-      type: object
-      properties:
-        successMessage:
-          type: string
-    ErrorResponse:
-      type: object
-      properties:
-        errorMessage:
-          type: string
