--- src/test/resources/examples/differentErrorResponseBody/client/ApiService.kt
@@ -1,36 +1,0 @@
-package examples.differentErrorResponseBody.client
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import examples.differentErrorResponseBody.models.SuccessResponse
-import io.github.resilience4j.circuitbreaker.CircuitBreakerRegistry
-import okhttp3.OkHttpClient
-import kotlin.String
-import kotlin.Suppress
-import kotlin.collections.Map
-import kotlin.jvm.Throws
-
-/**
- * The circuit breaker registry should have the proper configuration to correctly action on circuit
- * breaker transitions based on the client exceptions [ApiClientException], [ApiServerException] and
- * [IOException].
- *
- * @see ApiClientException
- * @see ApiServerException
- */
-@Suppress("unused")
-class ExampleService(
-    private val circuitBreakerRegistry: CircuitBreakerRegistry,
-    objectMapper: ObjectMapper,
-    baseUrl: String,
-    client: OkHttpClient
-) {
-    var circuitBreakerName: String = "exampleClient"
-
-    private val apiClient: ExampleClient = ExampleClient(objectMapper, baseUrl, client)
-
-    @Throws(ApiException::class)
-    fun getExample(additionalHeaders: Map<String, String> = emptyMap()): ApiResponse<SuccessResponse> =
-        withCircuitBreaker(circuitBreakerRegistry, circuitBreakerName) {
-            apiClient.getExample(additionalHeaders)
-        }
-}
