--- src/test/resources/examples/okHttpClientMultiMediaType/models/Models.kt
@@ -1,201 +1,0 @@
-package examples.okHttpClientMultiMediaType.models
-
-import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.annotation.JsonSubTypes
-import com.fasterxml.jackson.annotation.JsonTypeInfo
-import com.fasterxml.jackson.annotation.JsonValue
-import java.time.OffsetDateTime
-import javax.validation.Valid
-import javax.validation.constraints.NotNull
-import javax.validation.constraints.Size
-import kotlin.Boolean
-import kotlin.Int
-import kotlin.String
-import kotlin.collections.List
-import kotlin.collections.Map
-
-data class AlternateResponseModel(
-    @param:JsonProperty("extra_first_attr")
-    @get:JsonProperty("extra_first_attr")
-    val extraFirstAttr: OffsetDateTime? = null,
-    @param:JsonProperty("extra_second_attr")
-    @get:JsonProperty("extra_second_attr")
-    val extraSecondAttr: Int? = null
-)
-
-@JsonTypeInfo(
-    use = JsonTypeInfo.Id.NAME,
-    include = JsonTypeInfo.As.EXISTING_PROPERTY,
-    property = "model_type",
-    visible = true
-)
-@JsonSubTypes(
-    JsonSubTypes.Type(
-        value = FirstModel::class,
-        name =
-        "first_model"
-    ),
-    JsonSubTypes.Type(
-        value = SecondModel::class,
-        name =
-        "second_model"
-    ),
-    JsonSubTypes.Type(value = ThirdModel::class, name = "third_model")
-)
-sealed class Content(
-    open val id: String? = null,
-    open val firstAttr: OffsetDateTime? = null,
-    open val secondAttr: String? = null,
-    open val thirdAttr: ContentThirdAttr? = null,
-    open val etag: String? = null
-) {
-    abstract val modelType: ContentModelType
-}
-
-enum class ContentModelType(
-    @JsonValue
-    val value: String
-) {
-    FIRST_MODEL("first_model"),
-
-    SECOND_MODEL("second_model"),
-
-    THIRD_MODEL("third_model");
-
-    companion object {
-        private val mapping: Map<String, ContentModelType> =
-            values().associateBy(ContentModelType::value)
-
-        fun fromValue(value: String): ContentModelType? = mapping[value]
-    }
-}
-
-enum class ContentThirdAttr(
-    @JsonValue
-    val value: String
-) {
-    ENUM_TYPE_1("enum_type_1"),
-
-    ENUM_TYPE_2("enum_type_2");
-
-    companion object {
-        private val mapping: Map<String, ContentThirdAttr> =
-            values().associateBy(ContentThirdAttr::value)
-
-        fun fromValue(value: String): ContentThirdAttr? = mapping[value]
-    }
-}
-
-enum class ContentType(
-    @JsonValue
-    val value: String
-) {
-    APPLICATION_JSON("application/json"),
-
-    APPLICATION_VND_CUSTOM_MEDIA_JSON("application/vnd.custom.media+json");
-
-    companion object {
-        private val mapping: Map<String, ContentType> = values().associateBy(ContentType::value)
-
-        fun fromValue(value: String): ContentType? = mapping[value]
-    }
-}
-
-data class FirstModel(
-    @param:JsonProperty("id")
-    @get:JsonProperty("id")
-    override val id: String? = null,
-    @param:JsonProperty("first_attr")
-    @get:JsonProperty("first_attr")
-    override val firstAttr: OffsetDateTime? = null,
-    @param:JsonProperty("second_attr")
-    @get:JsonProperty("second_attr")
-    override val secondAttr: String? = null,
-    @param:JsonProperty("third_attr")
-    @get:JsonProperty("third_attr")
-    override val thirdAttr: ContentThirdAttr? = null,
-    @param:JsonProperty("etag")
-    @get:JsonProperty("etag")
-    override val etag: String? = null,
-    @param:JsonProperty("extra_first_attr")
-    @get:JsonProperty("extra_first_attr")
-    val extraFirstAttr: List<String>? = null,
-    @get:JsonProperty("model_type")
-    @get:NotNull
-    @param:JsonProperty("model_type")
-    override val modelType: ContentModelType = ContentModelType.FIRST_MODEL
-) : Content(id, firstAttr, secondAttr, thirdAttr, etag)
-
-data class OtherQueryResult(
-    @param:JsonProperty("items")
-    @get:JsonProperty("items")
-    @get:NotNull
-    @get:Size(min = 0)
-    @get:Valid
-    val items: List<AlternateResponseModel>
-)
-
-data class QueryResult(
-    @param:JsonProperty("items")
-    @get:JsonProperty("items")
-    @get:NotNull
-    @get:Size(min = 0)
-    @get:Valid
-    val items: List<Content>
-)
-
-data class SecondModel(
-    @param:JsonProperty("id")
-    @get:JsonProperty("id")
-    override val id: String? = null,
-    @param:JsonProperty("first_attr")
-    @get:JsonProperty("first_attr")
-    override val firstAttr: OffsetDateTime? = null,
-    @param:JsonProperty("second_attr")
-    @get:JsonProperty("second_attr")
-    override val secondAttr: String? = null,
-    @param:JsonProperty("third_attr")
-    @get:JsonProperty("third_attr")
-    override val thirdAttr: ContentThirdAttr? = null,
-    @param:JsonProperty("etag")
-    @get:JsonProperty("etag")
-    override val etag: String? = null,
-    @param:JsonProperty("extra_first_attr")
-    @get:JsonProperty("extra_first_attr")
-    val extraFirstAttr: String? = null,
-    @param:JsonProperty("extra_second_attr")
-    @get:JsonProperty("extra_second_attr")
-    val extraSecondAttr: Boolean? = null,
-    @get:JsonProperty("model_type")
-    @get:NotNull
-    @param:JsonProperty("model_type")
-    override val modelType: ContentModelType = ContentModelType.SECOND_MODEL
-) : Content(id, firstAttr, secondAttr, thirdAttr, etag)
-
-data class ThirdModel(
-    @param:JsonProperty("id")
-    @get:JsonProperty("id")
-    override val id: String? = null,
-    @param:JsonProperty("first_attr")
-    @get:JsonProperty("first_attr")
-    override val firstAttr: OffsetDateTime? = null,
-    @param:JsonProperty("second_attr")
-    @get:JsonProperty("second_attr")
-    override val secondAttr: String? = null,
-    @param:JsonProperty("third_attr")
-    @get:JsonProperty("third_attr")
-    override val thirdAttr: ContentThirdAttr? = null,
-    @param:JsonProperty("etag")
-    @get:JsonProperty("etag")
-    override val etag: String? = null,
-    @param:JsonProperty("extra_first_attr")
-    @get:JsonProperty("extra_first_attr")
-    val extraFirstAttr: OffsetDateTime? = null,
-    @param:JsonProperty("extra_second_attr")
-    @get:JsonProperty("extra_second_attr")
-    val extraSecondAttr: Int? = null,
-    @get:JsonProperty("model_type")
-    @get:NotNull
-    @param:JsonProperty("model_type")
-    override val modelType: ContentModelType = ContentModelType.THIRD_MODEL
-) : Content(id, firstAttr, secondAttr, thirdAttr, etag)
